Refactoring notifications
