+++ src/middleware.ts
@@ -14,8 +14,4 @@ const middleware = () => {
   Vue.use(VueKonva);
 };
 
-const a = {
-  installed: false,
-};
-
 export default middleware;
