--- src/notification/CssGroup.vue
@@ -1,11 +1,0 @@
-<template>
-  <transition-group :name="name">
-    <slot/>
-  </transition-group>
-</template>
-<script>
-export default {
-  name: 'CssGroup',
-  props: ['name']
-}
-</script>
